Extract subshell module
