+++ src/run/executable_call.rs
@@ -1,13 +1,6 @@
-use super::{exit_status_to_code, ExecutableArgs, ExecutablePath};
-use crate::cli;
-use crate::prelude::*;
-use std::ffi::OsString;
+use super::{ExecutableArgs, ExecutablePath};
 use std::fmt::{Display, Write};
-use std::io::{self, BufRead, BufReader, Read};
 use std::path::Path;
-use std::process::{self, Child, Command, ExitCode, Stdio};
-use std::sync::mpsc;
-use std::{env, thread};
 
 /// information to call an `App`s executable, as it is defined by the user
 #[derive(Clone)]
@@ -54,74 +47,6 @@ pub(crate) struct ExecutableCall {
 }
 
 impl ExecutableCall {
-  /// Executes the given executable with the given arguments.
-  /// The returned `ExitCode` also indicates failure if there has been any output.
-  #[allow(clippy::unwrap_used)]
-  pub(crate) fn check_output(&self, args: &[String], apps_to_include: &[ExecutableCall]) -> Result<ExitCode> {
-    let (sender, receiver) = mpsc::channel();
-    let mut cmd = Command::new(&self.executable_path);
-    cmd.args(&self.args);
-    cmd.args(args);
-    let mut paths_to_include = vec![self.executable_path.as_path().parent().unwrap()];
-    for app_to_include in apps_to_include {
-      paths_to_include.push(app_to_include.executable_path.as_path().parent().unwrap());
-    }
-    add_paths(&mut cmd, &paths_to_include);
-    cmd.stdout(Stdio::piped());
-    cmd.stderr(Stdio::piped());
-    let mut process = cmd.spawn().map_err(|err| UserError::CannotExecuteBinary {
-      call: self.format_with_extra_args(args),
-      reason: err.to_string(),
-    })?;
-    let Some(stdout) = process.stdout.take() else {
-      return Err(UserError::CannotOpenSubshellStream);
-    };
-    monitor_output(stdout, sender.clone());
-    let Some(stderr) = process.stderr.take() else {
-      return Err(UserError::CannotOpenSubshellStream);
-    };
-    monitor_output(stderr, sender.clone());
-    monitor_exit(process, sender);
-    let mut encountered_output = false;
-    let mut exit_code = ExitCode::SUCCESS;
-    let mut stdout = io::stdout();
-    for event in receiver {
-      match event {
-        Event::PermanentLine(line) | Event::TempLine(line) => {
-          encountered_output = true;
-          let mut colored_line: Vec<u8> = Vec::with_capacity(line.len() + BASH_RED.len() + BASH_CLEAR.len());
-          colored_line.extend(BASH_RED);
-          colored_line.extend(&line);
-          colored_line.extend(BASH_CLEAR);
-          if let Err(err) = io::Write::write_all(&mut stdout, &colored_line) {
-            eprintln!("Cannot print colored text: {err}");
-          }
-        }
-        Event::UnterminatedLine(line) => {
-          encountered_output = true;
-          let mut colored_line: Vec<u8> = Vec::with_capacity(line.len() + BASH_RED.len() + BASH_CLEAR.len() + 1);
-          colored_line.extend(BASH_RED);
-          colored_line.extend(&line);
-          colored_line.extend(BASH_CLEAR);
-          colored_line.push(b'\n');
-          if let Err(err) = io::Write::write_all(&mut stdout, &colored_line) {
-            eprintln!("Cannot print colored text: {err}");
-          }
-        }
-        Event::Ended { exit_status } => {
-          exit_code = exit_status_to_code(exit_status);
-          break;
-        }
-      }
-    }
-    if encountered_output {
-      let mut call = vec![self.executable_path.as_path().file_name().unwrap_or_default().to_string_lossy().to_string()];
-      call.extend(args.to_owned());
-      return Err(UserError::ProcessEmittedOutput { cmd: call.join(" ") });
-    }
-    Ok(exit_code)
-  }
-
   /// provides a printable version of this `ExecutableCall` when called with additional arguments
   pub(crate) fn format_with_extra_args(&self, args: &[String]) -> String {
     let mut result = self.to_string();
@@ -130,25 +55,6 @@ impl ExecutableCall {
       result.push_str(arg);
     }
     result
-  }
-
-  /// Runs the given executable with the given arguments.
-  /// Streams output to the user's terminal.
-  #[allow(clippy::unwrap_used)]
-  pub(crate) fn stream_output(&self, args: &[String], apps_to_include: &[ExecutableCall]) -> Result<ExitCode> {
-    let mut cmd = Command::new(&self.executable_path);
-    cmd.args(&self.args);
-    cmd.args(args);
-    let mut paths_to_include = vec![self.executable_path.as_path().parent().unwrap()];
-    for app_to_include in apps_to_include {
-      paths_to_include.push(app_to_include.executable_path.as_path().parent().unwrap());
-    }
-    add_paths(&mut cmd, &paths_to_include);
-    let exit_status = cmd.status().map_err(|err| UserError::CannotExecuteBinary {
-      call: self.format_with_extra_args(args),
-      reason: err.to_string(),
-    })?;
-    Ok(exit_status_to_code(exit_status))
   }
 }
 
@@ -161,91 +67,6 @@ impl Display for ExecutableCall {
     }
     Ok(())
   }
-}
-
-/// escape sequence to print red output on the shell
-const BASH_RED: &[u8] = "\x1B[0;31m".as_bytes();
-/// escape sequence to reset the output color on the shell
-const BASH_CLEAR: &[u8] = "\x1B[0m".as_bytes();
-
-/// events that can happen with subshells
-pub(crate) enum Event {
-  /// a line of output to STDOUT or STDERR terminated by LF
-  PermanentLine(Vec<u8>),
-  /// a line of output to STDOUT or STDERR terminated by CR
-  TempLine(Vec<u8>),
-  /// a line of output to STDOUT or STDERR without a CR or LF at the end
-  UnterminatedLine(Vec<u8>),
-  /// the process has ended with the given exit code
-  Ended { exit_status: process::ExitStatus },
-}
-
-/// adds the given dirs to the PATH env variable of the given cmd
-pub(crate) fn add_paths(cmd: &mut Command, dirs: &[&Path]) {
-  cmd.envs(env::vars_os());
-  let new_path = if let Some(mut path) = env::var_os("PATH") {
-    // PATH env var is set to something here, could be empty string
-    for dir in dirs {
-      if !path.is_empty() {
-        path.push(":");
-      }
-      path.push(dir.as_os_str());
-    }
-    path
-  } else {
-    // PATH env var is empty here
-    let mut path = OsString::new();
-    for dir in dirs {
-      if !path.is_empty() {
-        path.push(":");
-      }
-      path.push(dir);
-    }
-    path
-  };
-  cmd.env("PATH", new_path);
-}
-
-/// starts a thread that monitors the given STDOUT or STDERR stream
-fn monitor_output<R: 'static + Read + Send>(stream: R, sender: mpsc::Sender<Event>) {
-  let mut reader = BufReader::new(stream);
-  thread::spawn(move || loop {
-    let buffer = match reader.fill_buf() {
-      Ok(buffer) => buffer,
-      Err(err) => cli::exit(format!("cannot write subshell output into buffer: {err}")),
-    };
-    if buffer.is_empty() {
-      break;
-    }
-    let consumed = buffer.iter().take_while(|c| **c != b'\n' && **c != b'\x0D').count();
-    let total = if consumed < buffer.len() {
-      // stopped at one of the EOL characters
-      consumed + 1
-    } else {
-      // found no EOL character
-      consumed
-    };
-    let line = buffer[0..total].to_owned();
-    reader.consume(total);
-    let event = match line.get(consumed) {
-      Some(b'\n') => Event::PermanentLine(line),
-      Some(b'\x0D') => Event::TempLine(line),
-      _ => Event::UnterminatedLine(line),
-    };
-    if let Err(err) = sender.send(event) {
-      eprintln!("cannot send subshell output through internal pipe: {err}");
-    }
-  });
-}
-
-/// starts the thread that monitors for process exit
-fn monitor_exit(mut process: Child, sender: mpsc::Sender<Event>) {
-  thread::spawn(move || {
-    let exit_status = process.wait().unwrap_or_default();
-    if let Err(err) = sender.send(Event::Ended { exit_status }) {
-      cli::exit(format!("cannot send exit signal through internal pipe: {err}"));
-    }
-  });
 }
 
 #[cfg(test)]
@@ -265,6 +86,8 @@ mod tests {
     fn unix_success() {
       use std::io::Write;
       use std::os::unix::fs::PermissionsExt;
+
+      use crate::subshell;
       let tempdir = tempfile::tempdir().unwrap();
       let executable_path = tempdir.path().join("executable");
       let mut file = fs::File::create(&executable_path).unwrap();
@@ -277,7 +100,7 @@ mod tests {
         executable_path: ExecutablePath::from(executable_path),
         args: vec![],
       };
-      let have = executable_call.stream_output(&[], &[]).unwrap();
+      let have = subshell::stream_output(&executable_call, &[], &[]).unwrap();
       // HACK: is there a better way to compare ExitCode?
       assert_eq!(format!("{have:?}"), S("ExitCode(unix_exit_status(0))"));
     }
@@ -286,6 +109,7 @@ mod tests {
     #[cfg(unix)]
     fn unix_error() {
       use crate::filesystem::make_file_executable;
+      use crate::subshell;
       let tempdir = tempfile::tempdir().unwrap();
       let executable_path = tempdir.path().join("executable");
       fs::write(&executable_path, b"#!/bin/sh\nexit 3").unwrap();
@@ -294,7 +118,7 @@ mod tests {
         executable_path: ExecutablePath::from(executable_path),
         args: vec![],
       };
-      let have = executable_call.stream_output(&[], &[]).unwrap();
+      let have = subshell::stream_output(&executable_call, &[], &[]).unwrap();
       // HACK: is there a better way to compare ExitCode?
       assert_eq!(format!("{have:?}"), S("ExitCode(unix_exit_status(3))"));
     }
@@ -302,6 +126,8 @@ mod tests {
     #[test]
     #[cfg(windows)]
     fn windows_success() {
+      use crate::subshell;
+
       let tempdir = tempfile::tempdir().unwrap();
       let executable_path = tempdir.path().join("executable.cmd");
       fs::write(&executable_path, b"echo hello").unwrap();
@@ -309,7 +135,7 @@ mod tests {
         executable_path: ExecutablePath::from(executable_path),
         args: vec![],
       };
-      let have = executable_call.stream_output(&[], &[]).unwrap();
+      let have = subshell::stream_output(&executable_call, &[], &[]).unwrap();
       // HACK: is there a better way to compare ExitCode?
       assert_eq!(format!("{have:?}"), S("ExitCode(ExitCode(0))"));
     }
@@ -324,7 +150,7 @@ mod tests {
         executable_path: ExecutablePath::from(executable_path),
         args: vec![],
       };
-      let have = executable_call.stream_output(&[], &[]).unwrap();
+      let have = subshell::stream_output(&executable_call, &[], &[]).unwrap();
       // HACK: is there a better way to compare ExitCode?
       assert_eq!(format!("{have:?}"), S("ExitCode(ExitCode(3))"));
     }
